requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
deprecation==2.1.0
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.40.2
Flask==3.1.1
flask-cors==6.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
greenlet==3.2.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
realtime==2.7.0
redis==6.4.0
requests==2.32.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.1
StrEnum==0.4.15
stripe==12.4.0
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
